+++ packages/DataGrid/stories/helpers/data.integration.worker.js
@@ -1,10 +1,10 @@
-import nanoid from "nanoid";
+import { v4 as uuidv4 } from "uuid";
 import rawData from "./data.integration";
 
 export function setDataIds() {
   const newData = rawData.map(item => {
     return {
-      key: nanoid(),
+      key: uuidv4(),
       ...item,
     };
   });