+++ packages/Confirmation/CHANGELOG.md
@@ -11,3 +11,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - Updated: Created and placed all constants into type.js file. [@kaan.darcey](https://github.com/KDarcey).
+
+## [1.0.8-alpha.0] - 2021-02-01
+
+### Changed
+
+- Replace `nanoid` by `uuid`.